+++ storage/database_factory.py
@@ -1,13 +1,14 @@
 from peewee import MySQLDatabase
 from environment import Environment
+from discord.ext import tasks
 
 
 class DatabaseFactory:
     def __init__(self):
         env = Environment()
         self.config = env.get_config()
-
         self._database = None
+        self.stay_alive.start()
 
     def get_database_connection(self):
         if self._database is None:
@@ -17,7 +18,8 @@ class DatabaseFactory:
                                            password=self.config.get("Database", "Password")
                                            )
 
-        if self._database.is_closed():
-            self._database.connect()
-
         return self._database
+
+    @tasks.loop(seconds=30)
+    async def stay_alive(self):
+        self.get_database_connection().connection().ping(True)