+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Database
+- Refactored the database connection so that it pings to stay alive. 
 
 ## [4.1.0] - 2019-09-23
 ### Role Flairing